feat(rav_test): Organize better testing code
